Fix issue: click to chat message notification
