+++ src/utils/PushNotification-parse.ts
@@ -6,6 +6,7 @@ import { checkAndRemovePnTokenViaSip } from '../api/sip'
 import { accountStore } from '../stores/accountStore'
 import { getAuthStore } from '../stores/authStore'
 import { getCallStore } from '../stores/callStore'
+import { chatStore } from '../stores/chatStore'
 import { Nav } from '../stores/Nav'
 import { BrekekeUtils } from './RnNativeModules'
 import { toBoolean } from './string'
@@ -14,6 +15,7 @@ import { waitTimeout } from './waitTimeout'
 const keysInCustomNotification = [
   'title',
   'threadId',
+  'isGroupChat',
   'alert',
   'body',
   'message',
@@ -227,9 +229,16 @@ export const parse = async (raw: { [k: string]: unknown }, isLocal = false) => {
   } else if (isLocal) {
     // Chat local notification
     await signInByLocalNotification(n)
+
     if (n.threadId && n.threadId.length > 0) {
-      nav.customPageIndex = nav.goToPageChatDetail
-      waitTimeout().then(() => nav.goToPageChatDetail({ buddy: n.threadId }))
+      if (n?.isGroupChat) {
+        waitTimeout().then(() =>
+          chatStore.handleMoveToChatGroupDetail(n.threadId),
+        )
+      } else {
+        nav.customPageIndex = nav.goToPageChatDetail
+        waitTimeout().then(() => nav.goToPageChatDetail({ buddy: n.threadId }))
+      }
     } else {
       nav.customPageIndex = nav.goToPageChatRecents
       waitTimeout().then(nav.goToPageChatRecents)
@@ -305,6 +314,7 @@ export type ParsedPn = {
   title: string
   body: string
   threadId: string
+  isGroupChat?: boolean
   alert: string
   message: string
   from: string
